balthworker/requirements.txt
amqp==5.1.1
billiard==3.6.4.0
celery==5.2.7
cffi==1.15.1
charset-normalizer==3.0.1
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
cryptography==39.0.0
kombu==5.2.4
pdfminer.six==20221105
pdfrw==0.4
prompt-toolkit==3.0.36
pycparser==2.21
pytz==2022.7.1
six==1.16.0
vine==5.0.0
wcwidth==0.2.6
